+++ deltafi-ui/CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 ### Added
 - Loading indicator to DeltaFile Viewer
+- Flow Configuation page
 
 ## [0.15.0] - 2021-12-20
 ### Added